gapplicationimpl-dbus: Fix a -Wsign-conversion warning
These two variables are used for iterating over the children of a
`GVariant`, which is always done in terms of `size_t`. So they should be
`size_t`. Fixes a `-Wsign-conversion` warning.


Helps: #3405

+++ gio/gapplicationimpl-dbus.c
@@ -211,7 +211,7 @@ g_application_impl_method_call (GDBusConnection       *connection,
       const gchar *hint;
       GVariant *array;
       GFile **files;
-      gint n, i;
+      size_t n, i;
 
       flags = g_application_get_flags (impl->app);
       if ((flags & G_APPLICATION_HANDLES_OPEN) == 0)
